Move all sensors related code in `sensorslib`
This commit will create a new module library named `sensorslib` and will move all sensors related code in there; the moved files are `SensorsData`, `SensorsListener`, `SensorsData` and their tests.
The `sensors` package name inside `wear`'s module is renamed to `service` since now it only contains `SensorService`.
