+++ sensorslib/src/main/java/com/handmonitor/sensorslib/SensorsListener.kt
@@ -1,4 +1,4 @@
-package com.handmonitor.wear.sensors
+package com.handmonitor.sensorslib
 
 import android.content.Context
 import android.hardware.Sensor
