+++ sensorslib/src/test/java/com/handmonitor/sensorslib/SensorsConsumerTest.kt
@@ -1,4 +1,4 @@
-package com.handmonitor.wear.sensors
+package com.handmonitor.sensorslib
 
 import android.util.Log
 import com.google.common.truth.Truth.assertThat
