+++ sensorslib/src/test/java/com/handmonitor/sensorslib/SensorsListenerTest.kt
@@ -1,4 +1,4 @@
-package com.handmonitor.wear.sensors
+package com.handmonitor.sensorslib
 
 import android.content.Context
 import android.hardware.Sensor
