+++ settings.gradle
@@ -14,3 +14,4 @@ dependencyResolutionManagement {
 }
 rootProject.name = "HandMonitor"
 include ':wear'
+include ':sensorslib'
