+++ wear/build.gradle
@@ -63,6 +63,8 @@ dependencies {
 
     implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-android:1.6.4'
 
+    implementation(project(path: ":sensorslib"))
+
     implementation "org.tensorflow:tensorflow-lite:$tflite_version"
     implementation "org.tensorflow:tensorflow-lite-gpu:$tflite_version"
     implementation "org.tensorflow:tensorflow-lite-gpu-api:$tflite_version"
@@ -71,8 +73,8 @@ dependencies {
     kapt "androidx.room:room-compiler:$room_version"
     implementation "androidx.room:room-ktx:$room_version"
 
-    testImplementation "org.junit.jupiter:junit-jupiter-api:5.8.2"
-    testRuntimeOnly "org.junit.jupiter:junit-jupiter-engine:5.8.2"
+    testImplementation "org.junit.jupiter:junit-jupiter-api:$junit_version"
+    testRuntimeOnly "org.junit.jupiter:junit-jupiter-engine:$junit_version"
     testImplementation "io.mockk:mockk:1.13.2"
     testImplementation "com.google.truth:truth:1.1.3"
     testImplementation "androidx.room:room-testing:$room_version"
