+++ wear/src/main/java/com/handmonitor/wear/presentation/MainActivity.kt
@@ -24,7 +24,7 @@ import androidx.wear.compose.material.MaterialTheme
 import androidx.wear.compose.material.Text
 import com.handmonitor.wear.R
 import com.handmonitor.wear.presentation.theme.HandMonitorTheme
-import com.handmonitor.wear.sensors.SensorService
+import com.handmonitor.wear.service.SensorService
 
 class MainActivity : ComponentActivity() {
